Updated the UI
Reinvented the UI Containers, and UI Buttons.

Content Editable is now stored in an additional DIV Element to use as an anchor.

The BlueBook now sits off to the side.

But I thought the UI might be better with an absolute position so it doesn't effect the positions of the original DOM Tree Elements.

I think the Projects is setup to start working on the refactor.

also Fixed some bugs I left when I commented out a mystery conditional in the Merge_Previous function.