First Phase of Refactor
The Goal of the Refactor was to create Extra Files to separate the RichTextEditor, The Globals Variables, and the UI.

I also Renamed some files and variables to try and make it easier to understand.

Now the TEO-UI file is now the primary file used to create the Text Editor Object.

This Push is done to prepare for the next phase of the refactor.