Refactor Work and Links
Integrated the Traversal method into other functions that could make use of it.

Added a Backwards Traversal Method.

Added Bitwise Operators and methods to support Links and handle white spaces inherited from the html's file formatting.

Decoupled the UI and Global variables from the editor.

Recreated the Method to Target Trees when the selection is changed.

Phased out some of the CSS classes that toggled a single style and I am now applying that style directly.