Hot fixed Bug with selection.
The issue was 2 areas. HTML I wrote in by hand quick didn't follow the structure the editor was designed for.

The other issue occurred with what elements the the browser sets the selection to. Its not fully understood if my targeting system can settle on something other than a tree.